node-py-pycryptodome/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pycryptodome==3.20.0
pydantic==2.5.2
redis==5.0.1
